Get Intervention, Phase, and Status Info
Added three functions that take data as a JSON object and parse based on the first three requirements of the project. If you run this code currently, it parses the Hep A JSON file and prints 4 dictionaries: #1 is all the phases/frequencies, #2 is all the statuses/frequencies, #3 is all the intervention types/frequencies, and #4 is the number of interventions per trial/frequencies.